# perf: optimize custom bool unmarshal